tools/requirements.txt
pika
requests
python-dotenv
validators